feat(media): moderation path — pending thumbs with harvested facts, per-photo decisions, approval attaches
Photos add zero new moderation mechanics: the decision runs inside
ModerationService::decide()'s existing transaction, the per-photo choice rides
the existing decide POST, and the item-side record is the same change_history
row every other applied change gets.

Approving a submission approves its photos except the ones the curator
unticked; rejecting rejects all of them with no per-photo escape; needs-info
leaves them pending, because the rider is still being asked.

The curator judges a photo alongside the facts harvested from it — capture
month and how far the shot was from the pin — in one query per page rather than
one per row. takenAt is month-granular and the key is absent entirely when the
file carried no date.

+++ web/src/Media/PhotoProcessor.php
@@ -195,10 +195,12 @@ final class PhotoProcessor
             if (1 !== preg_match('#^(-?\d+)/(\d+)$#', $part, $matches) || '0' === $matches[2]) {
                 return null;
             }
-            $components[] = (int) $matches[1] / (int) $matches[2];
+            // Cast to float explicitly: an exact division of two ints is an int
+            // in PHP, and degrees/minutes/seconds must stay one numeric type.
+            $components[] = (float) ((int) $matches[1] / (int) $matches[2]);
         }
 
-        $degrees = $components[0] + $components[1] / 60 + $components[2] / 3600;
+        $degrees = $components[0] + $components[1] / 60.0 + $components[2] / 3600.0;
         if (null !== $ref && strtoupper($ref) === $negativeRef) {
             $degrees = -$degrees;
         }
